requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
aiokafka==0.11.0
boto3==1.35.24
redis==5.0.8
web3==7.2.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
motor==3.6.0
pytest==8.3.3
pytest-asyncio==0.24.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation-fastapi==0.61b0

httpx==0.27.2
